chatbot/pyproject.toml
[tool.poetry]
name = "ssu-chatbot"
version = "0.1.0"
description = ""
authors = ["Ian Carlson <carlsoni@sonoma.edu>"]
readme = "README.md"
packages = [
    { include = "chatbot", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.12"
async-typer = "^0.1.8"
beautifulsoup4 = "^4.12"
dependency-injector = {extras = ["yaml"], version = "^4.45.0"}
jinja2 = "^3.1"
langchain = "^0.3"
langchain-community = "^0.3"
langchain-huggingface = "^0.1.2"
langchain-openai = "^0.3"
langchain-weaviate = "^0.0.4"
langchainhub = "^0.1.21"
langserve = "^0.3"
pandas = "^2.2"
python-dotenv = "^1.0"
typer = "^0.9"
uvicorn = "^0.23"
weaviate-client = "^4.7"
langchain-text-splitters = "^0.3.6"

[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.35"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
chatbot-cli = "chatbot.cli.chatbot_cli:app"
